refactor: use mocked data & rename test suite's name
